Some more optmisations in getIcon
